Calendar Page In-Progress
Created calendar page that contains npm calendar and fetches expenses and incomes and displays them on the corresponding date based on their due_date. Calendar also formats how information is styled in calendar tiles based on income_or_expense and returns green text if income and red if expense. This is based on the last income_expense value on the current day so cannot format individual item in tile yet.